cauterize more stuff, start working on the Geometry Creator
