+++ include/nbl/asset/utils/CGeometryCreator.h
@@ -1,14 +1,13 @@
 // Copyright (C) 2018-2025 - DevSH Graphics Programming Sp. z O.O.
 // This file is part of the "Nabla Engine".
 // For conditions of distribution and use, see copyright notice in nabla.h
-#ifndef _NBL_ASSET_I_GEOMETRY_CREATOR_H_INCLUDED_
-#define _NBL_ASSET_I_GEOMETRY_CREATOR_H_INCLUDED_
+#ifndef _NBL_ASSET_C_GEOMETRY_CREATOR_H_INCLUDED_
+#define _NBL_ASSET_C_GEOMETRY_CREATOR_H_INCLUDED_
 
 
 #include "nbl/core/declarations.h"
 
-#include "nbl/asset/ICPUPolygonGeometry.h"
-#include "nbl/asset/utils/IMeshManipulator.h"
+#include "nbl/asset/utils/CPolygonGeometryManipulator.h"
 
 
 namespace nbl::asset
@@ -16,27 +15,17 @@ namespace nbl::asset
 
 //! Helper class for creating geometry on the fly.
 /** You can get an instance of this class through ISceneManager::getGeometryCreator() */
-class IGeometryCreator : public core::IReferenceCounted
+class CGeometryCreator final : public core::IReferenceCounted
 {
-		_NBL_INTERFACE_CHILD(IGeometryCreator) {}
-	public:
-		struct return_type
-		{
-			SVertexInputParams inputParams;
-			SPrimitiveAssemblyParams assemblyParams;
-			SBufferBinding<ICPUBuffer> bindings[ICPUMeshBuffer::MAX_ATTR_BUF_BINDING_COUNT];
-			SBufferBinding<ICPUBuffer> indexBuffer;
-			E_INDEX_TYPE indexType;
-			uint32_t indexCount;
-			core::aabbox3df bbox;
-		};
+		core::smart_refctd_ptr<CPolygonGeometryManipulator> m_defaultPolygonManipulator;
 
+	public:
 		//! Creates a simple cube mesh.
 		/**
 		\param size Dimensions of the cube.
 		\return Generated mesh.
 		*/
-		virtual return_type createCubeMesh(const core::vector3df& size=core::vector3df(5.f,5.f,5.f)) const =0;
+		core::smart_refctd_ptr<ICPUPolygonGeometry> createCube(const hlsl::float32_t3 size={5.f,5.f,5.f}) const;
 
 
 		//! Create an arrow mesh, composed of a cylinder and a cone.
@@ -53,11 +42,11 @@ class IGeometryCreator : public core::IReferenceCounted
 		\param colorCone color of the cone
 		\return Generated mesh.
 		*/
-		virtual return_type createArrowMesh(const uint32_t tesselationCylinder = 4,
+		core::smart_refctd_ptr<ICPUPolygonGeometry> createArrow(const uint32_t tesselationCylinder = 4,
 				const uint32_t tesselationCone = 8, const float height = 1.f,
 				const float cylinderHeight = 0.6f, const float widthCylinder = 0.05f,
 				const float widthCone = 0.3f, const video::SColor colorCylinder = 0xFFFFFFFF,
-				const video::SColor colorCone = 0xFFFFFFFF, IMeshManipulator* const meshManipulatorOverride = nullptr) const =0;
+				const video::SColor colorCone = 0xFFFFFFFF, IMeshManipulator* const meshManipulatorOverride = nullptr) const;
 
 
 		//! Create a sphere mesh.
@@ -67,8 +56,8 @@ class IGeometryCreator : public core::IReferenceCounted
 		\param polyCountY Number of quads used for the vertical tiling
 		\return Generated mesh.
 		*/
-		virtual return_type createSphereMesh(float radius = 5.f,
-				uint32_t polyCountX = 16, uint32_t polyCountY = 16, IMeshManipulator* const meshManipulatorOverride = nullptr) const =0;
+		core::smart_refctd_ptr<ICPUPolygonGeometry> createSphere(float radius = 5.f,
+				uint32_t polyCountX = 16, uint32_t polyCountY = 16, IMeshManipulator* const meshManipulatorOverride = nullptr) const;
 
 		//! Create a cylinder mesh.
 		/**
@@ -80,9 +69,9 @@ class IGeometryCreator : public core::IReferenceCounted
 		\param oblique (to be documented)
 		\return Generated mesh.
 		*/
-		virtual return_type createCylinderMesh(float radius, float length,
+		core::smart_refctd_ptr<ICPUPolygonGeometry> createCylinder(float radius, float length,
 				uint32_t tesselation,
-				const video::SColor& color=video::SColor(0xffffffff), IMeshManipulator* const meshManipulatorOverride = nullptr) const =0;
+				const video::SColor& color=video::SColor(0xffffffff), IMeshManipulator* const meshManipulatorOverride = nullptr) const;
 
 		//! Create a cone mesh.
 		/**
@@ -94,14 +83,14 @@ class IGeometryCreator : public core::IReferenceCounted
 		\param oblique (to be documented)
 		\return Generated mesh.
 		*/
-		virtual return_type createConeMesh(float radius, float length, uint32_t tesselation,
+		core::smart_refctd_ptr<ICPUPolygonGeometry> createCone(float radius, float length, uint32_t tesselation,
 				const video::SColor& colorTop=video::SColor(0xffffffff),
 				const video::SColor& colorBottom=video::SColor(0xffffffff),
-				float oblique=0.f, IMeshManipulator* const meshManipulatorOverride = nullptr) const =0;
+				float oblique=0.f, IMeshManipulator* const meshManipulatorOverride = nullptr) const;
 
-		virtual return_type createRectangleMesh(const core::vector2df_SIMD& size = core::vector2df_SIMD(0.5f, 0.5f)) const = 0;
+		core::smart_refctd_ptr<ICPUPolygonGeometry> createRectangle(const core::vector2df_SIMD& size = core::vector2df_SIMD(0.5f, 0.5f)) const;
 
-		virtual return_type createDiskMesh(float radius, uint32_t tesselation) const = 0;
+		core::smart_refctd_ptr<ICPUPolygonGeometry> createDisk(float radius, uint32_t tesselation) const;
 
 		//! Create a icosphere geometry
 		/**
@@ -110,7 +99,7 @@ class IGeometryCreator : public core::IReferenceCounted
 			\param smooth Specifies whether vertecies should be built for smooth or flat shading.
 		*/
 
-		virtual return_type createIcoSphere(float radius = 1.0f, uint32_t subdivision = 1, bool smooth = false) const = 0;
+		core::smart_refctd_ptr<ICPUPolygonGeometry> createIcoSphere(float radius = 1.0f, uint32_t subdivision = 1, bool smooth = false) const;
 
 };
